The halo elements are now stored _after_ the padding elements. This results in nrowshalo >= nrowspadded.
The remote column indices now have an additional offset and
context->hput_pos has been adjusted as well.

This addresses #205 but this ticket will stay open until further testing
has been done.

+++ include/ghost/machine.h
@@ -137,7 +137,7 @@ extern "C" {
      *
      * @return The SIMD width in bytes.
      */
-    size_t ghost_machine_simd_width();
+    int ghost_machine_simd_width();
 #ifdef __cplusplus
 }
 #endif
